cart delete route tested and works

+++ server/api/cart.js
@@ -1,6 +1,7 @@
-const router = require('express').Router()
+const express = require('express')
+const router = express.Router()
 const {Cart} = require('../db/models')
-
+router.use(express.json())
 // gets all producs in the cart for specified user:
 
 router.get('/', async (req, res) => {
@@ -23,17 +24,22 @@ router.get('/', async (req, res) => {
   }
 })
 
-// deletes item by id from the cart for specified user:
+// deletes product by id from the cart for specified user (which id is on the body of request):
 
 router.delete('/:id', (req, res, next) => {
-  Cart.destroy({
-    where: {
-      id: req.params.Id,
-      userId: req.params.userId
-    }
-  })
-    .then(() => res.status(204).end())
-    .catch(next)
+  const userId = req.body.userId
+  if (!userId) {
+    res.send('no user defined!')
+  } else {
+    Cart.destroy({
+      where: {
+        productId: req.params.id,
+        userId
+      }
+    })
+      .then(() => res.status(204).end())
+      .catch(next)
+  }
 })
 
 // creates new row in the cart with all params that's specified on req.body
